Feat/type inference (#14)
* fix: infer string type for identifier and date fields

* fix: widen generated integer fields to long

+++ generator/src/main/java/kr/go/h2spec/generator/emit/ClientEmitter.java
@@ -14,7 +14,7 @@ public class ClientEmitter {
     private static final String SERVICE_KEY_PARAM = "serviceKey";
     private static final Map<String, String> TYPE_MAP = Map.of(
             "string", "String",
-            "integer", "Integer",
+            "integer", "Long",
             "number", "Double",
             "boolean", "Boolean");
     private static final String DEFAULT_TYPE = "String";
